Use range-based for in more places + various fixes + switch to gmock
* differntiate logical and arithmetic 'not' operations during c-code generation
* following @turboencabulator 's example, removed old changelogs from files
* use cmake's external_project support to get gmock+gtest integrated into the build.
* various doxygen-related documentation fixes
* refactored UserProc dumps into debugPrintAll method
* moved methodss from sslparser.y to sslparser_support.cpp
* introduced llvm-like Const::get family of value builders
* converted tests to use gTest
* update x86 FCOM instruction semantics
* using in-class variable initialization in more places.
* updated regression tester to report crashes ( counts )

+++ include/constraint.h
@@ -12,12 +12,6 @@
  * OVERVIEW:   Definition of objects related to type constraints
  *============================================================================*/
 
-/*
- * $Revision$
- *
- * 22 Aug 03 - Mike: Created
- */
-
 #include "statement.h"
 #include "exphelp.h"
 #include <sstream>
@@ -77,7 +71,7 @@ typedef std::map<Exp*, LocationSet, lessExpStar>::iterator iterator;
     iterator begin() {return emap.begin();}
     iterator end()     {return emap.end();}
     void erase(iterator it) {emap.erase(it);}
-    int     size() {return emap.size();}
+    size_t size() {return emap.size();}
     // Add an equate (both ways)
     void addEquate(Exp* a, Exp* b) {
         emap[a].insert(b); emap[b].insert(a);}
@@ -108,9 +102,6 @@ public:
     void    substIntoDisjuncts(ConstraintMap& in);
     // Substitute the given constraintMap into the equates
     void    substIntoEquates(ConstraintMap& in);
-    // Perform "alpha substitution". Example:
-    //     <fixedtype*> = alphaX* and T[Y] = alphaX*    ->
-    //     T[Y] = <fixedtype*>
     void    alphaSubst();
 
     // Solve the constraints. If they can be solved, return true and put
